requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
pymongo==4.4.0
motor==3.2.0
python-dotenv==1.0.0
Werkzeug==2.3.6
dnspython==2.3.0
pandas==2.0.0
numpy==1.24.0
